Dodat interval i lastTimeExecuted i logika za interval

+++ rest-api/src/main/java/com/api/dto/JokeDto.java
@@ -47,11 +47,11 @@ public class JokeDto{
 		this.interval = interval;
 	}
 
-	public int getLastTimeExecuted() {
+	public long getLastTimeExecuted() {
 		return LastTimeExecuted;
 	}
 
-	public void setLastTimeExecuted(int lastTimeExecuted) {
+	public void setLastTimeExecuted(long lastTimeExecuted) {
 		LastTimeExecuted = lastTimeExecuted;
 	}
